v4.30

+++ User/motor/bsp_motor.h
@@ -56,7 +56,7 @@ void L_R_impulse(int speed, int j, int impulse);
 void R_R_impulse(int speed, int j, int impulse);
 void weitiao_2(void);
 
-void yaw_run(int16_t target_yaw, int16_t error_range) ;
+void yaw_run(float target_yaw, int16_t error_range) ;
 void weitiao(int gg);
 void oo(void);
 void se_huan(uint16_t num);
